+++ src/main/java/io/github/astrapi69/tree/api/ITree.java
@@ -24,25 +24,23 @@
  */
 package io.github.astrapi69.tree.api;
 
-import io.github.astrapi69.design.pattern.visitor.Acceptable;
-import io.github.astrapi69.design.pattern.visitor.Visitor;
-import io.github.astrapi69.tree.handler.ITreeNodeHandlerExtensions;
-import io.github.astrapi69.tree.handler.TreeNodeVisitorHandlerExtensions;
-import lombok.NonNull;
-
 import java.io.Serializable;
 import java.util.Collection;
 import java.util.List;
 
+import lombok.NonNull;
+import io.github.astrapi69.design.pattern.visitor.Acceptable;
+import io.github.astrapi69.design.pattern.visitor.Visitor;
+
 /**
- * The Interface {@link ITreeNode} holds the children in a {@link Collection} object
+ * The Interface {@link ITree} holds the children in a {@link Collection} object
  *
  * @param <V>
  *            the generic type of the value
  * @param <T>
  *            the generic type of the concrete tree node
  */
-public interface ITreeNode<V, T extends ITreeNode<V, T>> extends ITree<V, T>, Acceptable<Visitor<T>>
+public interface ITree<V, T> extends Serializable
 {
 
 	/**
@@ -51,10 +49,7 @@ public interface ITreeNode<V, T extends ITreeNode<V, T>> extends ITree<V, T>, Ac
 	 * @param child
 	 *            the child
 	 */
-	default void addChild(final T child)
-	{
-		ITreeNodeHandlerExtensions.addChild((T)this, child);
-	}
+	void addChild(final T child);
 
 	/**
 	 * Adds all the given children
@@ -62,10 +57,7 @@ public interface ITreeNode<V, T extends ITreeNode<V, T>> extends ITree<V, T>, Ac
 	 * @param children
 	 *            the children to add
 	 */
-	default void addChildren(final @NonNull Collection<T> children)
-	{
-		ITreeNodeHandlerExtensions.addChildren((T)this, children);
-	}
+	void addChildren(final @NonNull Collection<T> children);
 
 	/**
 	 * Returns all siblings of this node in the parent's children list. Returns null if this node is
@@ -73,10 +65,7 @@ public interface ITreeNode<V, T extends ITreeNode<V, T>> extends ITree<V, T>, Ac
 	 *
 	 * @return Returns all siblings of this node or null if this node is the root.
 	 */
-	default Collection<T> getAllSiblings()
-	{
-		return ITreeNodeHandlerExtensions.getAllSiblings((T)this);
-	}
+	Collection<T> getAllSiblings();
 
 	/**
 	 * Returns the previous sibling of this node in the parent's children list. Returns null if this
@@ -85,10 +74,7 @@ public interface ITreeNode<V, T extends ITreeNode<V, T>> extends ITree<V, T>, Ac
 	 * @return the next sibling of this node or null if this node is the root or is the parent's
 	 *         last child.
 	 */
-	default T getPreviousSibling()
-	{
-		return ITreeNodeHandlerExtensions.getPreviousSibling((T)this);
-	}
+	T getPreviousSibling();
 
 	/**
 	 * Returns the next sibling of this node in the parent's children list. Returns null if this
@@ -97,80 +83,131 @@ public interface ITreeNode<V, T extends ITreeNode<V, T>> extends ITree<V, T>, Ac
 	 * @return the next sibling of this node or null if this node is the root or is the parent's
 	 *         last child.
 	 */
-	default T getNextSibling()
-	{
-		return ITreeNodeHandlerExtensions.getNextSibling((T)this);
-	}
+	T getNextSibling();
 
 	/**
 	 * Gets the child count.
 	 *
 	 * @return the child count
 	 */
-	default int getChildCount()
-	{
-		return getChildren().size();
-	}
+	int getChildCount();
+
+	/**
+	 * Gets the children.
+	 *
+	 * @return the children
+	 */
+	Collection<T> getChildren();
+
+	/**
+	 * Sets the children.
+	 *
+	 * @param children
+	 *            the new children
+	 */
+	void setChildren(final Collection<T> children);
+
+	/**
+	 * Gets the optional display value.
+	 *
+	 * @return the display value
+	 */
+	String getDisplayValue();
+
+	/**
+	 * Sets the optional display value.
+	 *
+	 * @param displayValue
+	 *            the new optional display value
+	 */
+	void setDisplayValue(final String displayValue);
 
 	/**
 	 * Returns the distance from the root to this node. Returns 0 if this node is the root.
 	 *
 	 * @return the level from this node.
 	 */
-	default int getLevel()
-	{
-		return ITreeNodeHandlerExtensions.getLevel((T)this);
-	}
+	int getLevel();
+
+	/**
+	 * Gets the parent.
+	 *
+	 * @return the parent
+	 */
+	T getParent();
+
+	/**
+	 * Sets the parent.
+	 *
+	 * @param parent
+	 *            the new parent
+	 */
+	void setParent(final T parent);
 
 	/**
 	 * Gets the root object
 	 *
 	 * @return the root object
 	 */
-	default T getRoot()
-	{
-		return ITreeNodeHandlerExtensions.getRoot((T)this);
-	}
+	T getRoot();
+
+	/**
+	 * Gets the value.
+	 *
+	 * @return the value
+	 */
+	V getValue();
+
+	/**
+	 * Sets the value.
+	 *
+	 * @param value
+	 *            the new value
+	 */
+	void setValue(final V value);
 
 	/**
 	 * Checks for children.
 	 *
 	 * @return true, if successful
 	 */
-	default boolean hasChildren()
-	{
-		return ITreeNodeHandlerExtensions.hasChildren((T)this);
-	}
+	boolean hasChildren();
 
 	/**
 	 * Checks for parent
 	 *
 	 * @return true, if successful
 	 */
-	default boolean hasParent()
-	{
-		return ITreeNodeHandlerExtensions.hasParent((T)this);
-	}
+	boolean hasParent();
+
+	/**
+	 * Checks if is leaf.
+	 *
+	 * @return true, if is leaf
+	 */
+	boolean isLeaf();
+
+	/**
+	 * Sets the flag that indicates if this tree node is a node or a leaf
+	 *
+	 * @param leaf
+	 *            The flag to set that indicates if this tree node is a node or a leaf
+	 */
+	void setLeaf(boolean leaf);
 
 	/**
 	 * Checks if is node.
 	 *
 	 * @return true, if is node
 	 */
-	default boolean isNode()
-	{
-		return !isLeaf();
-	}
+	boolean isNode();
 
 	/**
-	 * Checks if this {@link ITreeNode} is the root {@link ITreeNode} object
+	 * Checks if this {@link ITree} is the root {@link ITree} object
 	 *
-	 * @return true, if this {@link ITreeNode} is the root {@link ITreeNode} object
+	 * @return true, if this {@link ITree} is the root {@link ITree} object
 	 */
-	default boolean isRoot()
-	{
-		return ITreeNodeHandlerExtensions.isRoot((T)this);
-	}
+	boolean isRoot();
 
 	/**
 	 * Removes the child.
@@ -178,34 +215,22 @@ public interface ITreeNode<V, T extends ITreeNode<V, T>> extends ITree<V, T>, Ac
 	 * @param child
 	 *            the child
 	 */
-	default void removeChild(final T child)
-	{
-		ITreeNodeHandlerExtensions.removeChild((T)this, child);
-	}
+	void removeChild(final T child);
 
 	/**
 	 * Removes all the children
 	 */
-	default void clearChildren()
-	{
-		ITreeNodeHandlerExtensions.clearChildren((T)this);
-	}
+	void clearChildren();
 
 	/**
 	 * Removes all the descendants
 	 */
-	default void clearAll()
-	{
-		ITreeNodeHandlerExtensions.clearAll((T)this);
-	}
+	void clearAll();
 
 	/**
 	 * Removes all the children
 	 */
-	default void removeChildren()
-	{
-		ITreeNodeHandlerExtensions.removeChildren((T)this);
-	}
+	void removeChildren();
 
 	/**
 	 * Removes all the given children
@@ -213,81 +238,55 @@ public interface ITreeNode<V, T extends ITreeNode<V, T>> extends ITree<V, T>, Ac
 	 * @param children
 	 *            the children to remove
 	 */
-	default void removeChildren(final @NonNull Collection<T> children)
-	{
-		ITreeNodeHandlerExtensions.removeChildren((T)this, children);
-	}
+	void removeChildren(final @NonNull Collection<T> children);
 
 	/**
-	 * {@inheritDoc}
-	 */
-	default void accept(final @NonNull Visitor<T> visitor)
-	{
-		TreeNodeVisitorHandlerExtensions.accept((T)this, visitor);
-	}
-
-	/**
-	 * Find all {@link ITreeNode} objects that have the same value as the given value
+	 * Find all {@link ITree} objects that have the same value as the given value
 	 *
 	 * @param value
 	 *            the value for the search process
 	 * @return a {@link Collection} object with all found occurrences that have the same value as
 	 *         the given value
 	 */
-	default Collection<T> findAllByValue(final V value)
-	{
-		return ITreeNodeHandlerExtensions.findAllByValue((T)this, value);
-	}
+	Collection<T> findAllByValue(final V value);
 
 	/**
-	 * Find all {@link ITreeNode} objects that have the same value as the given value
+	 * Find all {@link ITree} objects that have the same value as the given value
 	 *
 	 * @param value
 	 *            the value for the search process
 	 * @return a {@link Collection} object with all found occurrences that have the same value as
 	 *         the given value
 	 */
-	default T findByValue(final @NonNull V value)
-	{
-		return ITreeNodeHandlerExtensions.findByValue((T)this, value);
-	}
+	T findByValue(final @NonNull V value);
 
 	/**
-	 * Checks if the given {@link ITreeNode} object is a descendant of this tree node
+	 * Checks if the given {@link ITree} object is a descendant of this tree node
 	 *
 	 * @param treeNode
 	 *            the tree node to check
-	 * @return true if the given {@link ITreeNode} object is a descendant of this tree node
-	 *         otherwise false
+	 * @return true if the given {@link ITree} object is a descendant of this tree node otherwise
+	 *         false
 	 */
-	default boolean contains(T treeNode)
-	{
-		return ITreeNodeHandlerExtensions.contains((T)this, treeNode);
-	}
+	boolean contains(T treeNode);
 
 	/**
-	 * Checks if the given {@link Collection} object of {@link ITreeNode} objects are descendants of
+	 * Checks if the given {@link Collection} object of {@link ITree} objects are descendants of
 	 * this tree node
 	 *
 	 * @param treeNodes
 	 *            the children to add
-	 * @return true if the given {@link Collection} object of {@link ITreeNode} objects are
-	 *         descendants of this tree node otherwise false
+	 * @return true if the given {@link Collection} object of {@link ITree} objects are descendants
+	 *         of this tree node otherwise false
 	 */
-	default boolean containsAll(final @NonNull Collection<T> treeNodes)
-	{
-		return ITreeNodeHandlerExtensions.containsAll((T)this, treeNodes);
-	}
+	boolean containsAll(final @NonNull Collection<T> treeNodes);
 
 	/**
 	 * Traverse this node and adds all descendant with this included in to a {@link List} object
 	 *
 	 * @return a {@link List} object with this node and add all descendant
 	 */
-	default List<T> toList()
-	{
-		return ITreeNodeHandlerExtensions.toList((T)this);
-	}
+	List<T> toList();
 
 	/**
 	 * Traverse this node and adds all descendant with this included in to a {@link Collection}
@@ -295,9 +294,6 @@ public interface ITreeNode<V, T extends ITreeNode<V, T>> extends ITree<V, T>, Ac
 	 *
 	 * @return a {@link Collection} object with this node and add all descendant
 	 */
-	default Collection<T> traverse()
-	{
-		return TreeNodeVisitorHandlerExtensions.traverse((T)this);
-	}
+	Collection<T> traverse();
 
 }